work_list/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
djangorestframework-simplejwt
mysqlclient == 2.1.1